requirements.txt
atomicwrites==1.4.0
attrs==21.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
Django==2.2.6
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
